refactor: 패키지 및 VO 구조 변경
