+++ src/main/java/roomescape/dao/reservation/ReservationVo.java
@@ -1,0 +1,4 @@
+package roomescape.dao.reservation;
+
+public record ReservationVo(Long reservationId, String name, String date, Long timeId, String timeValue) {
+}
